requirements.txt
asgiref==3.4.1
Django==3.2.7
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.12.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
